# code/storage/app/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kanban_project_backend_api",
        "contact": {
            "name": "kent vejrup madsen",
            "url": "https://github.com/KentVejrupMadsen",
            "email": "kent.vejrup.madsen@protonmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/KentVejrupMadsen/Kanban-Project-Backend/blob/main/License.md"
        },
        "version": "1.0.0-alpha"
    },
    "servers": [
        {
            "url": "https://kanban.goalpioneers.com/api/1.0.0",
            "description": "public available server"
        },
        {
            "url": "http://localhost:8000/api/1.0.0",
            "description": "local development server"
        }
    ],
    "paths": {
...
        "/api/1.0.0/accounts/information/read": {
            "get": {
                "tags": [
                    "1.0.0",
                    "account",
                    "account-additional"
                ],
                "operationId": "ea305e1cc3e03cdcdb90ca069c2614c6",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/1.0.0/accounts/information/create": {
            "post": {
                "tags": [
                    "1.0.0",
                    "account",
                    "account-additional"
                ],
                "operationId": "49ebaed678ed332bbec20c5deef04e1e",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
        "/api/1.0.0/accounts/information/update": {
            "patch": {
                "tags": [
                    "1.0.0",
                    "account",
                    "account-additional"
                ],
                "operationId": "7e48e391ba10dd50c952622e9ca231a5",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
        "/api/1.0.0/accounts/information/delete": {
            "delete": {
                "tags": [
                    "1.0.0",
                    "account",
                    "account-additional"
                ],
                "operationId": "f9fc08be44f6d51f141efe15823ac082",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "account_logout": {
                "type": "http",
                "name": "authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            },
            "account_update": {
                "type": "http",
                "name": "authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            },
            "account_deletion": {
                "type": "http",
                "description": "",
                "name": "authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}